+++ R/bq-table.R
@@ -125,7 +125,12 @@ bq_table_delete <- function(x) {
 #' @rdname api-table
 #' @inheritParams bq_perform_copy
 #' @param dest Source and destination [bq_table]s.
-bq_table_copy <- function(x, dest, ..., quiet = NA) {
+bq_table_copy <- function(
+  x,
+  dest,
+  ...,
+  quiet = getOption("bigrquery.quiet", NA)
+) {
   x <- as_bq_table(x)
   dest <- as_bq_table(dest)
 
@@ -138,7 +143,7 @@ bq_table_copy <- function(x, dest, ..., quiet = NA) {
 #' @export
 #' @rdname api-table
 #' @inheritParams api-perform
-bq_table_upload <- function(x, values, ..., quiet = NA, json_digits = NA) {
+bq_table_upload <- function(x, values, ..., quiet = getOption("bigrquery.quiet", NA), json_digits = NA) {
   x <- as_bq_table(x)
 
   json_digits <- check_digits(json_digits)
@@ -150,7 +155,12 @@ bq_table_upload <- function(x, values, ..., quiet = NA, json_digits = NA) {
 
 #' @export
 #' @rdname api-table
-bq_table_save <- function(x, destination_uris, ..., quiet = NA) {
+bq_table_save <- function(
+  x,
+  destination_uris,
+  ...,
+  quiet = getOption("bigrquery.quiet", NA)
+) {
   x <- as_bq_table(x)
 
   job <- bq_perform_extract(x, destination_uris = destination_uris, ...)
@@ -161,7 +171,7 @@ bq_table_save <- function(x, destination_uris, ..., quiet = NA) {
 
 #' @export
 #' @rdname api-table
-bq_table_load <- function(x, source_uris, ..., quiet = NA, json_digits = NA) {
+bq_table_load <- function(x, source_uris, ..., quiet = getOption("bigrquery.quiet", NA), json_digits = NA) {
   x <- as_bq_table(x)
 
   json_digits <- check_digits(json_digits)